requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
datasets==3.1.0
dill==0.3.8
filelock==3.16.1
fonttools==4.37.2
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.26.5
idna==3.4
kiwisolver==1.4.4
kornia==0.7.3
kornia_rs==0.1.7
lpips==0.1.4
matplotlib==3.6.0
multidict==6.1.0
multiprocess==0.70.16
numpy==1.23.3
packaging==21.3
pandas==2.0.3
pillow==10.4.0
propcache==0.2.0
psutil==6.1.0
pyarrow==17.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2024.2
PyYAML==6.0
requests==2.32.3
safetensors==0.4.5
scipy==1.9.1
six==1.16.0
torch==1.11.0+cu113
torch-fidelity==0.3.0
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.67.1
typing_extensions==4.3.0
tzdata==2024.2
urllib3==1.26.12
xxhash==3.5.0
yarl==1.15.2
